requirements.txt
matplotlib==3.1.1
numpy==1.19.5
opencv-python==4.5.2.52
Pillow==8.4.0
tensorflow==2.6.2